Chore: Migrate REST API - registerPushToken to Typescript (#3902)
